Initial support for requiring a selected game group to navigate the app.
Created a specific controller for game groups (GameGroupPortal).
Moved the gameGroups action there from Application.
Created new interceptor that checks for a selected gamegroup in the
current session scope. If none is found it redirects the user to the
gameGroups route.
Printed a warning message in the gameGroups view when no groups
are found.